Fixed issues with for loop
The issue was within Environment::ancestor(), which was returning a clone of the environment, which is either just the current environment (if no enclosing environment exists) or the enclosing environment. Since the ancestor is mutated (e.g. fn assign_at()), it has to be on that instance of the Environment, not on a clone, as the changes aren't "saved".